Redirect to the path given in URL
This implements data-8/connector-instructors#16:

Right now, hitting a url like:

    https://data8.berkeley.edu/hub/interact?repo=data8assets&path=labs/lab01

Results in the student being redirected to the root of the repo:
`data8assets`.

Instead, it'd be more convenient if the student were redirected to the
path given in the URL: `data8assets/labs/lab01`.

Since you can specify multiple values for `path` in the URL, eg.

    https://data8.berkeley.edu/hub/interact?repo=data8assets&path=labs/lab01&path=labs/lab01/lab01.ipynb

The last path will be chosen to redirect to. In the above example, the
student will be sent directly to the notebook.

Resolves data-8/connector-instructors#16